# api_docs.json
{"openapi":"3.0.1","info":{"title":"è®¢åå¯è§åå¹³å° API ææ¡£","description":"## è®¢åå¯è§åæ°å­åç®¡çå¹³å°\n\n### ææ¯æ \n- Spring Boot 3.2.x\n- Java 21\n- MySQL 8.0+\n- MyBatis Plus\n- Knife4j (OpenAPI 3.0)\n\n### è®¤è¯è¯´æ\né¤ç»å½/æ³¨åæ¥å£å¤ï¼æææ¥å£é½éè¦å¨è¯·æ±å¤´ä¸­æºå¸¦ Tokenï¼\n\n```\nAuthorization: Bearer {token}\n```\n\n### éè¯¯ç è¯´æ\n- 200: æä½æå\n- 401: æªç»å½æ Token è¿æ\n- 403: æ æéè®¿é®\n- 500: æå¡å¨åé¨éè¯¯\n","contact":{"name":"å¼åç»","email":"dev@example.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8081","description":"æ¬å°å¼åç¯å¢"},{"url":"/","description":"å½åæå¡å¨"}],"security":[{"Bearer Authentication":[]}],"tags":[{"name":"ç¨æ·ç®¡ç","description":"ç¨æ·CRUDãç¶æç®¡çç­æ¥å£"},{"name":"è®¤è¯ç»å½","description":"ç¨æ·ç»å½ãç»åºãTokenå·æ°ç­è®¤è¯ç¸å³æ¥å£"}],"paths":{"/api/users/{id}":{"get":{"tags":["ç¨æ·ç®¡ç"],"summary":"æ¥è¯¢ç¨æ·è¯¦æ","description":"æ ¹æ®IDæ¥è¯¢ç¨æ·å®æ´ä¿¡æ¯","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserPageVO"}}}}}},"put":{"tags":["ç¨æ·ç®¡ç"],"summary":"æ´æ°ç¨æ·","description":"æ´æ°ç¨æ·ä¿¡æ¯ï¼é¨åå­æ®µï¼","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}},"delete":{"tags":["ç¨æ·ç®¡ç"],"summary":"å é¤ç¨æ·","description":"é»è¾å é¤ç¨æ·ï¼ä¸è½å é¤èªå·±","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users":{"post":{"tags":["ç¨æ·ç®¡ç"],"summary":"åå»ºç¨æ·","description":"åå»ºæ°ç¨æ·ï¼å¯ç èªå¨å å¯","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultLong"}}}}}}},"/api/auth/reset-password/{id}":{"post":{"tags":["è®¤è¯ç»å½"],"summary":"éç½®å¯ç ","description":"ç®¡çåéç½®ç¨æ·å¯ç ","operationId":"resetPassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/api/auth/refresh":{"post":{"tags":["è®¤è¯ç»å½"],"summary":"å·æ°Token","description":"ä½¿ç¨æ§Tokenæ¢åæ°Tokenï¼éè¦ Authorization: Bearer æ ¼å¼ï¼","operationId":"refreshToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/api/auth/logout":{"post":{"tags":["è®¤è¯ç»å½"],"summary":"ç¨æ·ç»åº","description":"ç¨æ·éåºç»å½","operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/auth/login":{"post":{"tags":["è®¤è¯ç»å½"],"summary":"ç¨æ·ç»å½","description":"æ¯æç¨æ·å/é®ç®±/ææºå·ç»å½","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultLoginVO"}}}}}}},"/api/auth/change-password":{"post":{"tags":["è®¤è¯ç»å½"],"summary":"ä¿®æ¹å¯ç ","description":"ç¨æ·ä¿®æ¹èªå·±çå¯ç ","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/{id}/unlock":{"patch":{"tags":["ç¨æ·ç®¡ç"],"summary":"è§£éç¨æ·","description":"è§£éç¨æ·è´¦æ·","operationId":"unlockUser","parameters":[{"name":"id","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/{id}/status":{"patch":{"tags":["ç¨æ·ç®¡ç"],"summary":"å¯ç¨/ç¦ç¨ç¨æ·","description":"åæ¢ç¨æ·å¯ç¨ç¶æï¼ä¸è½ç¦ç¨èªå·±","operationId":"updateUserStatus","parameters":[{"name":"id","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"isEnabled","in":"query","description":"æ¯å¦å¯ç¨ï¼0-ç¦ç¨ï¼1-å¯ç¨","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/{id}/lock":{"patch":{"tags":["ç¨æ·ç®¡ç"],"summary":"éå®ç¨æ·","description":"éå®ç¨æ·è´¦æ·","operationId":"lockUser","parameters":[{"name":"id","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"reason","in":"query","description":"éå®åå ","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/api/users/page":{"get":{"tags":["ç¨æ·ç®¡ç"],"summary":"åé¡µæ¥è¯¢ç¨æ·","description":"æ¯æå¤æ¡ä»¶ç­éãåé¡µãæåº","operationId":"queryUserPage","parameters":[{"name":"queryDTO","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserQueryDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultPageUserPageVO"}}}}}}},"/api/auth/current":{"get":{"tags":["è®¤è¯ç»å½"],"summary":"è·åå½åç¨æ·ä¿¡æ¯","description":"æ ¹æ®Tokenè·åå½åç»å½ç¨æ·ä¿¡æ¯","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultObject"}}}}}}}},"components":{"schemas":{"UserUpdateDTO":{"type":"object","properties":{"realName":{"maxLength":20,"minLength":2,"type":"string","description":"çå®å§å","example":"å¼ ä¸"},"email":{"type":"string","description":"é®ç®±","example":"zhangsan@example.com"},"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string","description":"ææºå·","example":"13800138000"},"departmentId":{"type":"integer","description":"é¨é¨ID","format":"int64","example":10},"departmentName":{"type":"string","description":"é¨é¨åç§°","example":"åä¸å¤§åº"},"position":{"type":"string","description":"èä½","example":"å®¢æ·ç»ç"},"employeeNo":{"type":"string","description":"å·¥å·","example":"EMP001"},"remark":{"type":"string","description":"ç¨æ·å¤æ³¨","example":"è´è´£åä¸åºåå®¢æ·ç®¡ç"},"isEnabled":{"type":"integer","description":"æ¯å¦å¯ç¨ï¼0-ç¦ç¨ï¼1-å¯ç¨","format":"int32","example":1},"avatar":{"type":"string","description":"å¤´åURL","example":"https://example.com/avatar.jpg"}},"description":"æ´æ°ç¨æ·è¯·æ±"},"ResultVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"},"timestamp":{"type":"integer","format":"int64"}}},"UserCreateDTO":{"required":["password","realName","username"],"type":"object","properties":{"username":{"maxLength":20,"minLength":2,"pattern":"^[a-zA-Z][a-zA-Z0-9_]{1,19}$","type":"string","description":"ç¨æ·å","example":"zhangsan"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"åå§å¯ç ","example":"123456"},"realName":{"maxLength":20,"minLength":2,"type":"string","description":"çå®å§å","example":"å¼ ä¸"},"email":{"type":"string","description":"é®ç®±","example":"zhangsan@example.com"},"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string","description":"ææºå·","example":"13800138000"},"departmentId":{"type":"integer","description":"é¨é¨ID","format":"int64","example":10},"departmentName":{"type":"string","description":"é¨é¨åç§°","example":"åä¸å¤§åº"},"position":{"type":"string","description":"èä½","example":"å®¢æ·ç»ç"},"employeeNo":{"type":"string","description":"å·¥å·","example":"EMP001"},"remark":{"type":"string","description":"ç¨æ·å¤æ³¨","example":"è´è´£åä¸åºåå®¢æ·ç®¡ç"}},"description":"åå»ºç¨æ·è¯·æ±"},"ResultLong":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"integer","format":"int64"},"timestamp":{"type":"integer","format":"int64"}}},"ResultString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"string"},"timestamp":{"type":"integer","format":"int64"}}},"LoginDTO":{"required":["account","password"],"type":"object","properties":{"account":{"type":"string","description":"ç»å½è´¦å·ï¼ç¨æ·å/é®ç®±/ææºå·ï¼","example":"zhangsan"},"password":{"type":"string","description":"å¯ç ","example":"123456"},"captcha":{"type":"string","description":"å¾å½¢éªè¯ç ","example":"abcd"},"captchaKey":{"type":"string","description":"éªè¯ç Key","example":"uuid-uuid-uuid-uuid"},"email":{"type":"boolean"},"phone":{"type":"boolean"},"username":{"type":"boolean"},"accountValid":{"type":"boolean"},"loginType":{"type":"string"}},"description":"ç»å½è¯·æ±"},"CurrentUserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"realName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"avatar":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"departmentId":{"type":"integer","format":"int64"},"departmentName":{"type":"string"},"userCode":{"type":"string"},"employeeNo":{"type":"string"},"position":{"type":"string"}},"description":"å½åç¨æ·ä¿¡æ¯"},"DataScopeInfo":{"type":"object","properties":{"type":{"type":"string","description":"æ°æ®æéç±»å","example":"SELF"},"typeCode":{"type":"integer","description":"æ°æ®æéç±»åä»£ç ","format":"int32","example":3},"departmentId":{"type":"integer","description":"é¨é¨ID","format":"int64","example":10},"departmentName":{"type":"string","description":"é¨é¨åç§°","example":"åä¸å¤§åº"},"description":{"type":"string","description":"æ°æ®æéæè¿°","example":"åªè½æ¥çèªå·±çæ°æ®"},"allDataScope":{"type":"boolean"},"selfDataScope":{"type":"boolean"},"departmentDataScope":{"type":"boolean"}},"description":"æ°æ®æéä¿¡æ¯"},"LoginVO":{"type":"object","properties":{"token":{"type":"string","description":"JWTè®¿é®ä»¤ç","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"tokenType":{"type":"string","description":"ä»¤çç±»å","example":"Bearer"},"expiresIn":{"type":"integer","description":"è¿ææ¶é´ï¼ç§ï¼","format":"int64","example":604800},"requireChangePassword":{"type":"boolean","description":"æ¯å¦éè¦ä¿®æ¹å¯ç ","example":false},"passwordExpireTime":{"type":"string","description":"å¯ç è¿ææ¶é´","format":"date-time"},"userInfo":{"$ref":"#/components/schemas/CurrentUserDTO"},"roles":{"type":"array","description":"ç¨æ·è§è²åè¡¨","example":["CUSTOMER_MANAGER"],"items":{"type":"string","description":"ç¨æ·è§è²åè¡¨","example":"[\"CUSTOMER_MANAGER\"]"}},"permissions":{"type":"array","description":"ç¨æ·æéåè¡¨","example":["ORDER:VIEW","ORDER:CREATE"],"items":{"type":"string","description":"ç¨æ·æéåè¡¨","example":"[\"ORDER:VIEW\",\"ORDER:CREATE\"]"}},"dataScope":{"$ref":"#/components/schemas/DataScopeInfo"},"systemAdmin":{"type":"boolean"},"dataAdmin":{"type":"boolean"}},"description":"ç»å½ååº"},"ResultLoginVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginVO"},"timestamp":{"type":"integer","format":"int64"}}},"ChangePasswordDTO":{"required":["confirmPassword","newPassword","oldPassword"],"type":"object","properties":{"oldPassword":{"maxLength":20,"minLength":6,"type":"string","description":"æ§å¯ç ","example":"123456"},"newPassword":{"maxLength":20,"minLength":6,"type":"string","description":"æ°å¯ç ","example":"Abc123!@#"},"confirmPassword":{"type":"string","description":"ç¡®è®¤å¯ç ","example":"Abc123!@#"},"passwordMatch":{"type":"boolean"},"newPasswordStrong":{"type":"boolean"},"newPasswordDifferent":{"type":"boolean"},"passwordStrengthDescription":{"type":"string"}},"description":"ä¿®æ¹å¯ç è¯·æ±"},"ResultUserPageVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserPageVO"},"timestamp":{"type":"integer","format":"int64"}}},"UserPageVO":{"type":"object","properties":{"id":{"type":"integer","description":"ç¨æ·ID","format":"int64","example":1},"username":{"type":"string","description":"ç¨æ·å","example":"zhangsan"},"userCode":{"type":"string","description":"ç¨æ·ç¼å·","example":"USER001"},"realName":{"type":"string","description":"çå®å§å","example":"å¼ ä¸"},"email":{"type":"string","description":"é®ç®±","example":"zhangsan@example.com"},"phone":{"type":"string","description":"ææºå·","example":"13800138000"},"avatar":{"type":"string","description":"å¤´åURL","example":"https://example.com/avatar.jpg"},"departmentId":{"type":"integer","description":"é¨é¨ID","format":"int64","example":10},"departmentName":{"type":"string","description":"é¨é¨åç§°","example":"åä¸å¤§åº"},"position":{"type":"string","description":"èä½","example":"å®¢æ·ç»ç"},"employeeNo":{"type":"string","description":"å·¥å·","example":"EMP001"},"isEnabled":{"type":"integer","description":"æ¯å¦å¯ç¨ï¼0-ç¦ç¨ï¼1-å¯ç¨","format":"int32","example":1},"isLocked":{"type":"integer","description":"æ¯å¦éå®ï¼0-æªéå®ï¼1-å·²éå®","format":"int32","example":0},"lockedTime":{"type":"string","description":"éå®æ¶é´","format":"date-time"},"lockedReason":{"type":"string","description":"éå®åå ","example":"å¯ç éè¯¯æ¬¡æ°è¿å¤"},"lastLoginTime":{"type":"string","description":"æåç»å½æ¶é´","format":"date-time"},"lastLoginIp":{"type":"string","description":"æåç»å½IP","example":"192.168.1.100"},"loginCount":{"type":"integer","description":"ç»å½æ¬¡æ°","format":"int32","example":100},"roles":{"type":"array","description":"ç¨æ·è§è²åè¡¨","example":["CUSTOMER_MANAGER"],"items":{"type":"string","description":"ç¨æ·è§è²åè¡¨","example":"[\"CUSTOMER_MANAGER\"]"}},"roleNames":{"type":"array","description":"ç¨æ·è§è²åç§°åè¡¨","example":["å®¢æ·ç»ç"],"items":{"type":"string","description":"ç¨æ·è§è²åç§°åè¡¨","example":"[\"å®¢æ·ç»ç\"]"}},"createdAt":{"type":"string","description":"åå»ºæ¶é´","format":"date-time"},"createdBy":{"type":"integer","description":"åå»ºäººID","format":"int64","example":1},"createdByUsername":{"type":"string","description":"åå»ºäººç¨æ·å","example":"admin"},"updatedAt":{"type":"string","description":"æ´æ°æ¶é´","format":"date-time"},"remark":{"type":"string","description":"ç¨æ·å¤æ³¨","example":"è´è´£åä¸åºåå®¢æ·ç®¡ç"},"enabled":{"type":"boolean"},"lockedStatus":{"type":"boolean"}},"description":"ç¨æ·åé¡µååº"},"UserQueryDTO":{"type":"object","properties":{"current":{"type":"integer","description":"å½åé¡µç ","format":"int64","example":1},"size":{"type":"integer","description":"æ¯é¡µå¤§å°","format":"int64","example":10},"sortField":{"type":"string","description":"æåºå­æ®µ","example":"created_at"},"sortOrder":{"type":"string","description":"æåºæ¹å","example":"DESC"},"username":{"type":"string","description":"ç¨æ·åï¼æ¨¡ç³æ¥è¯¢ï¼","example":"zhang"},"realName":{"type":"string","description":"çå®å§åï¼æ¨¡ç³æ¥è¯¢ï¼","example":"å¼ "},"email":{"type":"string","description":"é®ç®±ï¼æ¨¡ç³æ¥è¯¢ï¼","example":"example"},"phone":{"type":"string","description":"ææºå·ï¼ç²¾ç¡®æ¥è¯¢ï¼","example":"13800138000"},"departmentId":{"type":"integer","description":"é¨é¨ID","format":"int64","example":10},"isEnabled":{"type":"integer","description":"æ¯å¦å¯ç¨ï¼0-ç¦ç¨ï¼1-å¯ç¨","format":"int32","example":1},"isLocked":{"type":"integer","description":"æ¯å¦éå®ï¼0-æªéå®ï¼1-å·²éå®","format":"int32","example":0}},"description":"ç¨æ·æ¥è¯¢è¯·æ±"},"OrderItem":{"type":"object","properties":{"column":{"type":"string"},"asc":{"type":"boolean"}}},"PageUserPageVO":{"type":"object","properties":{"records":{"type":"array","items":{"$ref":"#/components/schemas/UserPageVO"}},"total":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int64"},"current":{"type":"integer","format":"int64"},"orders":{"type":"array","writeOnly":true,"items":{"$ref":"#/components/schemas/OrderItem"}},"optimizeCountSql":{"$ref":"#/components/schemas/PageUserPageVO"},"searchCount":{"$ref":"#/components/schemas/PageUserPageVO"},"optimizeJoinOfCountSql":{"type":"boolean","writeOnly":true},"maxLimit":{"type":"integer","format":"int64","writeOnly":true},"countId":{"type":"string","writeOnly":true},"pages":{"type":"integer","format":"int64"}}},"ResultPageUserPageVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageUserPageVO"},"timestamp":{"type":"integer","format":"int64"}}},"ResultObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"},"timestamp":{"type":"integer","format":"int64"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","description":"è¯·è¾å¥ JWT Tokenï¼æ ¼å¼ï¼Bearer {token}","scheme":"bearer","bearerFormat":"JWT"}}},"x-openapi":{"x-setting":{"customCode":200,"language":"zh-CN","enableSwaggerModels":true,"swaggerModelName":"å®ä½ç±»åè¡¨","enableReloadCacheParameter":false,"enableAfterScript":true,"enableDocumentManage":true,"enableVersion":false,"enableRequestCache":true,"enableFilterMultipartApis":false,"enableFilterMultipartApiMethodType":"POST","enableHost":false,"enableHostText":"","enableDynamicParameter":false,"enableDebug":true,"enableFooter":true,"enableFooterCustom":false,"enableSearch":true,"enableOpenApi":true,"enableHomeCustom":false,"enableGroup":true,"enableResponseCode":true},"x-markdownFiles":[]}}